✨ feat: 실시간 amr 지나온 경로, 남은 경로 추가

+++ be/smartfactory/src/main/java/info/smartfactory/domain/bottleneck/dto/request/AddBottleneckRequest.java
@@ -19,7 +19,7 @@ public class AddBottleneckRequest {
 
     private Long missionId;
 
-    private Long bottleneckPeriod;
+    private Integer bottleneckPeriod;
 
     private LocalDateTime bottleneckTime;
